Move Retention app config to Retention context
