+++ src/Contexts/Retention/Campaign/infrastructure/config/config.ts
@@ -13,7 +13,7 @@ const convictConfig = convict({
       doc: 'The Mongo connection URL',
       format: String,
       env: 'MONGO_URL',
-      default: 'mongodb://localhost:27017/mooc-backend-dev'
+      default: 'mongodb://localhost:27017/retention'
     }
   }
 });
