+++ src/apps/retention/dependency-injection/application.yaml
