+++ src/apps/retention/dependency-injection/application_staging.yaml
